requirements.txt
flask
flask-sqlalchemy
psycopg2-binary
marshmallow
flask-migrate
azure-ai-textanalytics==5.1.0
python-dotenv
gunicorn